browserbase/pyproject.toml
[tool.poetry]
name = "browserbase-client"
version = "0.1.0"
description = ""
authors = ["Xavier <xavierh93@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12"
playwright = ">=1.52.0,<2.0.0"
browserbase = "^1.4.0"
python-dotenv = "^1.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
